rename outwatch.dom.* to outwatch.*
