+++ bench/src/main/scala/ChildrenPerformanceBenchmark.scala
@@ -2,8 +2,8 @@ package outwatch
 
 import cats.effect.SyncIO
 
-import outwatch.dom._
-import outwatch.dom.dsl._
+import outwatch._
+import outwatch.dsl._
 // import outwatch.reactive.handler._
 import colibri.ext.monix._
 import outwatch.ext.monix.handler._
