+++ bench/src/main/scala/VNodeConstructionBenchmark.scala
@@ -1,11 +1,11 @@
 package outwatch
 
-import outwatch.dom._
-import outwatch.dom.dsl._
+import outwatch._
+import outwatch.dsl._
 import monix.execution.ExecutionModel.SynchronousExecution
 import monix.execution.schedulers.TrampolineScheduler
 import monix.execution.Scheduler
-import outwatch.dom.interpreter.SnabbdomOps
+import outwatch.interpreter.SnabbdomOps
 
 import org.scalajs.dom.{ document, window }
 import scala.scalajs.js
