+++ tests/src/test/scala/outwatch/ScenarioTestSpec.scala
@@ -6,8 +6,7 @@ import monix.reactive.{Observable, Observer}
 import org.scalatest.Assertion
 import outwatch.Deprecated.IgnoreWarnings.initEvent
 
-import outwatch.dom._
-import outwatch.dom.dsl._
+import outwatch.dsl._
 import colibri.ext.monix._
 import outwatch.ext.monix.handler._
 import outwatch.util._
